requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
numpy
llama-cpp-python
speechrecognition
pyaudio
